refactor: move page components into separate folder
